fix(deps): use correct 'cache' commit hash
The `cypress-io/github-actions-cache` commit hash was different in the
`package-lock.json` file and the `package.json` file. I merely changed
the hash listed in `package.json` to match that listed in
`package-lock.json`.

This also fixes an error where the `lib/` directory wasn't built in the
latest `cypress-io/github-actions-create` commit.