fix: respect the CYPRESS_CACHE_FOLDER environment variable (#253)
When using `install: false`, the Cypress binary needs to be installed manually. If this is done a separate job, the caching gets complicated if the binary is in the home directory. Setting the env var `CYPRESS_CACHE_FOLDER: ${{ github.workspace }}/node_modules/.cache/Cypress` makes this trivial, but is currently not supported by this action as the cache folder path is hard-coded. This change adds support for overriding the path with the environment var.