chore: use command vs. command-prefix for yarn pnp suggestions
This is required because using `yarn` command-prefix will end up running: 
`npx yarn` which doesnt resolve correctly in pnp
and results in using yarn v1 not the specified yarn. 

Using `command` will overwrite the entire command and allow `yarn dlx` 
to be used instead of `npx` which properly runs and resolves.